+++ code/produto-service/src/main/java/com/alissonrsousa/produtoservice/repository/FornecedorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.alissonrsousa.produtoservice.model.Fornecedor;
 
-public interface FornecedorRepository extends CrudRepository<Fornecedor, Long> {
+public interface FornecedorRepository extends CrudRepository<Fornecedor, Integer> {
 
 }
